Modify SolutionTimeAdaptiveDT to make dt consistent accross all processors.
This update is to prevent the possibility of processors becoming out of sync
while computing the size of the next time step with the SolutionTimeAdaptiveDT
TimeStepper. Previously, the ratio of solution time to time step size was based
on local solution time for each processor. These ratios are used to determine
whether to increase or decrease dt for the next timestep. Because the ratios
differ across processors, it is possible for one processor to recommend
increasing dt while another processor recommends decreasing dt. To prevent this
behavior, this commit introduces the use of the maximum solution time over all
processors to compute the (solution time / dt) ratio, resulting in all
processors using the same ratio to compute the next dt. This results in each
processors recommending the same dt for the next time step. Closes #13426.